Upgrade nodegit; Pin yarn/node; Fix install bug;
Upgraded nodegit, because node-gyp@<7.x relies on python2, which is not installed by default on MacOS anymore (in fact, it's difficult to install period).

Pinned yarn and node using Volta.

Using the git:// protocol for vim-fugitive resulted in a timeout. Must be a bug in the new version of nodegit. Switching to the URL fixed it.